ksdflksa

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->decimal('salary')->nullable();
             $table->time('check_in_time')->nullable();
             $table->time('check_out_time')->nullable();
-            $table->decimal('office_time', 4,2)->nullable();
+            $table->integer('office_time')->nullable();
             $table->unsignedBigInteger('office_id')->nullable();
             $table->foreign('office_id')->references('id')->on('users')->onDelete('cascade');
             $table->rememberToken();
